+++ packages/hooks/src/index.ts
@@ -4,3 +4,5 @@ export * from './usePortal';
 export * from './useIntersection';
 export * from './useResize';
 export * from './useStatic';
+export * from './useDrag';
+export * from './useDrop';