+++ packages/web/select/src/newSelect.tsx
@@ -3,7 +3,15 @@ import Select from 'react-select';
 import { SelectProps, IOptions } from '.';
 
 export const NewSelect = (props: SelectProps) => {
-  const { options, onChange, multiple } = props;
+  const {
+    options,
+    onChange,
+    multiple,
+    disabled,
+    placeholder,
+    defaultValue,
+    styles,
+  } = props;
 
   const handleChange = (option: any) => {
     if (multiple) {
@@ -20,7 +28,12 @@ export const NewSelect = (props: SelectProps) => {
       options={options}
       onChange={handleChange}
       classNamePrefix="demon"
+      className="demon-select"
       isMulti={multiple}
+      isDisabled={disabled}
+      placeholder={placeholder}
+      styles={styles}
+      defaultValue={defaultValue}
     />
   );
 };