+++ packages/web/select/stories/Select.stories.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Select } from '../src';
 
 export default {
-  title: 'Components/base/select',
+  title: 'Components/Form/Select',
   component: Select,
 };
 
@@ -60,6 +60,7 @@ export const SelectStory = () => {
         isNew
         style={{ width: 300 }}
       />
+      <div style={{ height: 10, width: 10 }}></div>
       <Select
         options={options}
         onChange={(value, option) => {